requirements.txt
alembic==1.16.1
blinker==1.9.0
click==8.2.1
Flask==3.0.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.3
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
packaging==25.0
psycopg2-binary==2.9.10
python-dotenv==1.0.0
SQLAlchemy==2.0.41
typing-extensions==4.14.0
watchdog
Werkzeug==3.1.3
discord.py
requests
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
pytest
pytest-mock
PyYAML
ratelimit
backoff
SQLAlchemy-serializer==1.4.1
google-generativeai
aiohttp
